rename database -> user
