+++ index.js
@@ -9,7 +9,7 @@ const ejs= require('ejs');
 const { check } = require('express-validator/check');
 const { validationResult } = require('express-validator/check');
 const sendgridtransport = require('nodemailer-sendgrid-transport');
-const {User} = require('./database');
+const {User} = require('./user');
 const passport = require('passport');
 const LocalStrategy = require('passport-local').Strategy;
 const { initializingPassport } = require('./config/passport-local-strategy');
